+++ cmd/tools.go
@@ -12,4 +12,6 @@ import (
 
 	_ "github.com/ory/go-acc"
 	_ "github.com/ory/x/tools/listx"
+
+	_ "github.com/gobuffalo/packr/v2/..."
 )