+++ pipeline/authz/keto_engine_acp_ory_test.go
@@ -28,7 +28,7 @@ import (
 	"net/url"
 	"testing"
 
-	"github.com/spf13/viper"
+	"github.com/ory/viper"
 
 	"github.com/ory/oathkeeper/driver/configuration"
 	"github.com/ory/oathkeeper/internal"