+++ test/e2e/run.sh
@@ -16,7 +16,7 @@ export OATHKEEPER_PROXY=http://127.0.0.1:6660
 export OATHKEEPER_API=http://127.0.0.1:6661
 export GO111MODULE=on
 
-go install github.com/ory/oathkeeper
+(cd ../../; make install)
 go install github.com/ory/oathkeeper/test/e2e/okapi
 go install github.com/ory/oathkeeper/test/e2e/okclient